Harden the simulator membership bloom filters
The FastFilter and BloomFilter cast the expected-insertion count to an
int for the backing array, silently wrapping/saturating past 2^31; use
Math.toIntExact so an oversized filter fails loudly instead. Drop three
-v2 fast-filter type names from reference.conf that no longer exist in
fastfilter 1.0.2 (valueOf throws if selected).

The BloomFilter probes a fixed four hashes to reuse the sketch's, so
its "optimal" sizing was a misnomer; rename it and document that the
false-positive rate is approximate and leans small by design.

+++ simulator/src/main/java/com/github/benmanes/caffeine/cache/simulator/membership/bloom/FastFilter.java
@@ -40,7 +40,7 @@ public final class FastFilter implements Membership {
 
   public FastFilter(Config config) {
     var settings = new FastFilterSettings(config);
-    keys = new long[(int) settings.membership().expectedInsertions()];
+    keys = new long[Math.toIntExact(settings.membership().expectedInsertions())];
     filterType = settings.filterType();
     bitsPerKey = settings.bitsPerKey();
     reset();
